pyproject.toml
[project]
name = "vision3d"
version = "0.1.0"
description = "Open-source 3D object detection framework for autonomous driving and robotics"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
    # Deep Learning
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    # Training Engine
    "pytorch-lightning>=2.2.0",
    # Configuration
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    # Data I/O
    "Pillow>=10.0.0",
    "numpy>=1.26.0",
    "scipy>=1.12.0",        # Hungarian algorithm (scipy.optimize.linear_sum_assignment)
    # Visualisation (Foxglove MCAP)
    "mcap>=1.1.0",
    "foxglove-schemas-protobuf>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.3.0",
    "mypy>=1.9.0",
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]
# Dataset conversion — install when running scripts/convert_nuscenes.py
nuscenes = [
    "nuscenes-devkit>=1.1.11",
]

[project.scripts]
train = "vision3d.train:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vision3d"]

# ---------------------------------------------------------------------------
# uv
# ---------------------------------------------------------------------------
[tool.uv]
dev-dependencies = [
    "ruff>=0.3.0",
    "mypy>=1.9.0",
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

# ---------------------------------------------------------------------------
# Ruff
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "PT"]
ignore = ["E501"]

# ---------------------------------------------------------------------------
# Mypy
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "."]
addopts = "-v --tb=short"
